BIS-38/SP-210: CRC32 checksum feeder fails on sprint screening server Adapt unit test.
SVN: 26157

+++ common/sourceTest/java/ch/systemsx/cisd/common/io/hierarchical_content/HDF5ContainerBasedHierarchicalContentNodeTest.java
@@ -102,7 +102,7 @@ public class HDF5ContainerBasedHierarchicalContentNodeTest extends AssertJUnit
             {
                 builder.append("\n");
             }
-            builder.append(Long.toHexString(node.getChecksumCRC32()));
+            builder.append(Integer.toHexString(node.getChecksumCRC32()));
             builder.append(" ").append(node.getFileLength());
             builder.append(" ").append(node.getName());
         }
